config.py protect used configs
Previously deleting a config would delete all associated results_files.
This is not what we want, and somewhat dangertous since a couple wrong
clicks could lead to massive data loss.
An error message is now displayed when attempting to remove a config
which has results_files associated.

+++ server/web2py/applications/vidjil/controllers/config.py
@@ -115,15 +115,19 @@ def confirm():
 
 def delete():
     if (auth.can_modify_config(request.vars['id'])):
-        #delete results_file using this config
-        db(db.results_file.config_id==request.vars["id"]).delete()
+        count = db(db.results_file.config_id==request.vars["id"]).count()
 
-        #delete config
-        db(db.config.id==request.vars["id"]).delete()
+        if (count == 0):
+            #delete config
+            db(db.config.id==request.vars["id"]).delete()
 
-        res = {"redirect": "config/index",
-               "message": "config '%s' deleted" % request.vars["id"]}
-        log.admin(res)
+            res = {"redirect": "config/index",
+                   "message": "config '%s' deleted" % request.vars["id"]}
+            log.admin(res)
+        else:
+            res = {"redirect": "config/index",
+                    "success": "false",
+                    "message": "cannot delete a config that has been used"}
         return gluon.contrib.simplejson.dumps(res, separators=(',',':'))
     return error_message(ACCESS_DENIED)
